# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Katalog Produk Global API",
        "description": "Dokumentasi API untuk Modul 1",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/internal/catalog/attributes": {
            "post": {
                "tags": [
                    "Module 1 - Catalog Metadata"
                ],
                "summary": "Create new attribute",
                "operationId": "storeAttribute",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "color"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Warna"
                                    },
                                    "data_type": {
                                        "type": "string",
                                        "example": "select"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
...
        "/api/v1/internal/catalog/products/{id}/attributes/sync": {
            "put": {
                "tags": [
                    "Module 1 - Catalog Metadata"
                ],
                "summary": "Sync product inherited attributes",
                "description": "Sync attributes for a product based on its primary category hierarchy",
                "operationId": "syncProductAttributes",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Synced successfully"
                    }
                }
            }
        },
        "/api/v1/internal/catalog/categories/{id}/attributes/sync": {
            "post": {
                "tags": [
                    "Module 1 - Catalog Metadata"
                ],
                "summary": "Sync category attribute definitions",
                "description": "Attach attribute definitions to a category and trigger product re-syncs",
                "operationId": "syncCategoryAttributes",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "attributes": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "attribute_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "is_required": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Synced successfully"
                    }
                }
            }
        },
        "/api/v1/internal/catalog/products": {
            "post": {
                "tags": [
                    "Module 1 - Catalog Metadata"
                ],
                "summary": "Create new product",
                "operationId": "storeProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "sku": {
                                        "type": "string",
                                        "example": "SKU-123"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "product-123"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Product 123"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 15000
                                    },
                                    "categories": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "category_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "is_primary": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/v1/internal/catalog/products/{id}": {
            "put": {
                "tags": [
                    "Module 1 - Catalog Metadata"
                ],
                "summary": "Update product",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Product"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                }
            }
        }
    },
...
}